Correct windows error
MeshMath.cxx 
Define the fonction round 
Change variable name

ExternalVTK.cmake
Upload vtk version 
Add a patch for vtk : VTKPatch.cmake

ParametricMeshToSPHARMSpatialObjectFilter.cxx
Problem with function pow . Pow need a double instead a int

SphericalHarmonicPolynomial.txx
Change the definition of the function SetCoefs

Applications/CMakeLists.txt
Change the verion of ImageMath

ParticleModuleParameters.h and ParticleModuleComputation.h
Delete the <dirent.h> , no use in the two function

ShapeAnalysisModuleComputation.h , Parameters.cxx , Parameters.h and ShapeAnalysisModuleComputation.cxx
Delete the <dirent.h> , 
Use itksys:: SystemTools and itksys::Directory instead of dirent.h 

spharm-pdm.cmake
f2c.lib : Add an option for Windows

itkMesh3DProcrustesAlignFilter.h
Define M_PI

+++ Libraries/Shape/SpatialObject/SphericalHarmonicSpatialObject.cxx
@@ -71,7 +71,7 @@ void SphericalHarmonicSpatialObject::SetCoefs(SphericalHarmonicSpatialObject::Co
 {
   m_Coefs = coeflist;
   m_Count = m_Coefs.size(); // Each item in the vector is a 3D point.
-  m_Harmonic = (int) floor(sqrt(m_Count) ) - 1;
+  m_Harmonic = (int) floor(sqrt(double(m_Count)) ) - 1;
   // std::cout<<"m_Count = " <<m_Count <<std::endl;
   ComputeHiddenMeshSpatialObject();
 }
